Remove unnecessary `echo` calls
They were redundant. I can't think of a succinct way to explain why
this is a thing, though.

BTW, I also removed the leading zeros within the ANSI color escape
sequences, because they were redundant. Such a sequence is primarily
(if not entirely) split up into a series of what you could call
'commands', separated by semi-colons (';').

Below are various ways to adjust the primary attribute of the text, not
so much about the color of the foreground or background, with the
exception of '7', which reverses these colors.

0          = Reset
1          = Bright
2          = Dim
3          = Italic
4          = Underline
5          = Blink
7          = Reversed
9          = Strikethrough

You then have the colors, be they foreground or background, IE:

30         = Black
31         = Red
32         = Green
90         = Light Black
91         = Light Red
92         = Light Green

These are of course for the typical terminal type most of us all use.
I'm not familiar with other terminals types, so I cannot speak for
those.

So, with all this in mind, you can combine attributes with colors, IE:

1;31       = Bright; Red
2;95       = Dim; Light Blue
1;30       = Bright; Black

Hope this helps!