Rework & optimize `OS` variable functionality
I've left in a simpler(-ish) version using sed(1), in-case it's
preferred.

The loop method uses a standard `while read` loop I've talked about
almost religiously on the 'Learn Linux' channel, so feel free to resort
to that to see what I did and why I did it.

To summarize, however:

I've read each line of the file using the `read` builtin and by
redirecting the contents of the file into the loop (into `read`), using
the `<` redirection operator.

I've set an input field separator (IFS) for only this instance of
`read`, in order to delimit each field by a '=' instead of the default
of ' ' (whitespace).

By providing both `KEY` and `VALUE`, I've told `read` to assign the
first field (what is before the '=') to the variable called `KEY`, and
to assign anything thereafter (after '=') to the `VALUE` variable.

The `if` statement tests for whether the `KEY` is equal to the value of
'NAME' — in other words, whether we're on the right line. If the
condition is true, we're on the right line, so grab the `VALUE`.

Regarding the sed(1) method, I should point out that the method of
`/MATCH/!d` means lines containing 'MATCH' (REGEX) will be deleted. By
using `!`, the use of `d` (delete) is inverted; negated. It's possible
to instead use `/MATCH/p` if you're using the `-n` flag with sed(1),
but that's just more to type and a pain to remember, so it's not my
preferred approach.

The reason I used `break` in the loop is to avoid unnecessary
processing, which makes things more efficient, since at that point, we
have the data we want, so iterating over the proceeding lines is
pointless and wasteful; this is especially important with large amounts
of data.