Rework and greatly optimize getting RAM values
I was going to go with a simpler approach, but it just wasn't doing it
for me, so I went with my usual `while read` loop. I also made use of
the `printf` builtin to add commas, which I thought you might like.
When you have a lot of RAM (IIRC, you have 64GB), the numbers get
awkward to read.

This approach completely removes the need for `echo` (which was
redundant*, BTW), two AWK calls, and two grep(1) calls (both of which
were made redundant by the use of AWK).

For future reference, you can do as grep(1) was doing, in AWK, by, IE:

```
awk '/MemAvailable/ {print($2)}' /proc/meminfo
```

Furthermore, I recommend that you try to make your REGEX more robust, -
because you or others may (will likely) match unwanted values. The
above demonstrates an inaccurate REGEX, although in this case, we know
it's likely not going to be an issue, but with other data, it
definitely can be.

For example, if a line had 'MemAvailable_Stuff', then it too would be
matched. A more accurate REGEX would be

```
awk '/^MemAvailable:/ {print($2)}' /proc/meminfo
```

This matters because your code can be rendered incredibly buggy and
unreliable if a REGEX (for example) is not so robust.

* The reason the `echo` calls were redundant is because the result of
  the result was being saved into the variable, instead of just the
  result; hope that makes some sense.