Add TODO for final list of `echo` calls
It's almost right, but it needs to be reworked. This is a pain in the
arse to achieve; I usually do this sort of thing in Perl, but I have
done this in various shell projects, such as AutoExec.

You'd need to get the length of each value to be displayed, then use
the widths with something like `printf`, because of its `*` (IE: `%*s`)
functionality, allowing you to set the column width/spacing
dynamically. You could use AWK, but it'd probably be easier and it
would certainly be more efficient to use the shell itself.

I recommend looking at AutoExec. Also, have a look at Shlides. Although
I've somewhat abandoned Shlides*, it's still going to be useful, with
regards to having a layout and maintaining it. Be prepared to scratch
your brain and do some math. Honestly, I'd revisit this in the
future, when you've gotten a bit further in your adventures. It was a
long time before I started doing these sort of things, because they can
be very complicated.

* Terminal can be limited, and it takes such a lot of work to maintain
  a layout, especially in shell, that I didn't have the willpower to
  pursue it.